complete and test check restaurant owner identity for delete rest operation

+++ controllers/restaurant.js
@@ -78,12 +78,17 @@ async function editRestaurant(req, res) {
 }
 
 async function deleteRestaurant(req, res) {
-  try {
-    await modelRestaurant.deleteRestaurant(req.params.restId);
-    res.json("data has been deleted.");
-    // res.redirect('/');
-  } catch (err) {
-    console.error(err);
-    res.status(500).json({ errorMsg: err.message });
+  const restdata = await modelRestaurant.getRestaurantById(req.params.restId);
+  if (!restdata.owner || restdata.owner !== req.user.id) {
+    return res.status(401).json("Unauthorized");
+  } else {
+    try {
+      await modelRestaurant.deleteRestaurant(req.params.restId);
+      res.json("data has been deleted.");
+      // res.redirect('/');
+    } catch (err) {
+      console.error(err);
+      res.status(500).json({ errorMsg: err.message });
+    }
   }
 }
